modelstore/requirements.txt
Django==1.6.5
argparse==1.2.1
distribute==0.6.24
django-tastypie==0.11.1
django-transaction-hooks==0.1.4
python-dateutil==2.2
python-mimeparse==0.1.4
pyzmq==14.3.1
six==1.7.3
wsgiref==0.1.2
